pyproject.toml
[tool.poetry]
name = "python-benchmark"
version = "0.1.0"
description = ""
authors = ["Tamir Bahar <tbahar@vdoo.com>"]

[tool.poetry.dependencies]
python = "^3.10"
pytest-benchmark = "^3.4.1"
rich = "^12.0.1"
attrs = "^21.4.0"
cattrs = "^1.10.0"
typer = "^0.4.0"
click = "<8.1.0"
Jinja2 = "^3.1.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
